DMP-2626 moved ARM batch process response flow to a template class to allow reuse for DETS
